final structure of the app
