+++ server.js
@@ -4,11 +4,8 @@
  * This way, we have two different ways to use this module
  */
 const _ = require('lodash');
-const fs = require('fs');
-const http = require('http');
-const https = require('https');
 const express = require('express');
-const server = require('./src');
+const { appGenerator, serverGenerator } = require('./src');
 // Export
 module.exports = function(options = {}) {
   // We always overwrite it here to disable feature that shouldn't be use
@@ -20,29 +17,11 @@ module.exports = function(options = {}) {
   };
   options = _.merge(options, disable);
   // Generate the app
-  const { app, config, mockServerInstance } = server(options);
+  const { app, config, mockServerInstance } = appGenerator(options);
   // Static serving
   app.use(express.static(config.path, config.staticOptions));
   // Configure the server
-  let webserver;
-  if (config.https) {
-    let opts;
-    if (config.https.pfx) {
-      opts = {
-        pfx: fs.readFileSync(config.https.pfx),
-        passphrase: config.https.passphrase
-      };
-    } else {
-      opts = {
-        key: fs.readFileSync(config.https.key || config.devKeyPem),
-        cert: fs.readFileSync(config.https.cert || config.devCrtPem)
-      };
-    }
-    webserver = https.createServer(opts, app);
-  } else {
-    webserver = http.createServer(app);
-  }
-  webserver.listen(config.port, config.host, config.callback);
+  let webserver = serverGenerator(app, config);
   webserver.on('close', () => {
     mockServerInstance.close();
     // DebuggerInstance.close();
